Crea CRUDS, implementa JWT, completa SQL de carga inicial. Genera DTOs para omitir data sensible en respuestas.

+++ src/main/java/spa/nuevo/desafiotecnico/repository/UsuarioRepository.java
@@ -2,7 +2,7 @@
  * @ Author: Eduardo 'Ph1L' Rodríguez Bahamonde
  * @ Create Time: 2024-01-11 19:07:00
  * @ Modified by: Eduardo 'Ph1L' Rodríguez Bahamonde
- * @ Modified time: 2024-01-11 20:33:35
+ * @ Modified time: 2024-01-12 19:11:49
  * @ Description:
  */
 
@@ -16,7 +16,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
 
     Optional<Usuario> findByUsername(String username);
